pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "travelops"
version = "0.1.0"
description = "TravelOps Agent Demo for tester training (plan, tools, re-plan, Langfuse trace)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.1.0",
    "openai>=1.0.0",
    "langfuse>=2.0.0",
    "openinference-instrumentation-openai-agents>=1.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "pydantic>=2.0.0",
    "requests>=2.28.0",
    "duckduckgo-search>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7", "pytest-asyncio>=0.21"]
ui = ["streamlit>=1.28.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: mark test as integration (needs OPENAI_API_KEY)",
    "failure_mode: mark test as failure-mode scenario",
]
testpaths = ["tests"]

[project.scripts]
travelops = "run_agent:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[dependency-groups]
dev = ["pytest>=7", "pytest-asyncio>=0.21"]
